requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
mcp>=1.0.0
pydantic>=2.0.0
uvicorn>=0.30.0
fastapi>=0.110.0